+++ accounting/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.2.0] accounting
 ### Releases
 - [NPM v0.2.0] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.2.0 - accounting
+
+## 2023-11-06 01:47:30
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.3.0] accounting
+### Releases
+- [NPM v0.3.0] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.3.0 - accounting